requirements.txt
pandas
qlalchemy 
psycopg2-binary